Message rewrite, SCRAM Auth
Lots of big breaking changes. Auth works with a simple query, and caching seems to be working again so this commit is a 'stable' version, but refactoring of message handling continues...

Big stuff is:
SCRAM authentication fully implemented, so we can now authenticate using a password legitimately, and handle a simple query message.
MessageReader/PGMessage to better handle messages, still in progress are helpers to write directly to ByteWrappers and avoid allocating more arrays